requirements.txt
llama-cpp-python==0.2.90
fastapi==0.111.0
uvicorn==0.30.1
huggingface-hub==0.23.4
pydantic==2.7.4
python-dotenv==1.0.1
